Fixed High Score Ordering Bug; broke ADJ-NA mode, though...
Fixed the way the program updates the high scores list, so that scores
are overwritten and stored in the correct order. However, the method
used to do so is HORRIBLY inefficient, I'm sure. So, I'll need to come
back to it at a later date.

Also, I broke ADJ-NA mode. Not sure when or how this happened; now that
I think about it, it may have been a problem for the last couple of
updates. I'll have to look into it, and fix it later. Details on the
problem will be in the Issues page of GitHub.